My Chess Engine/requirements.txt
flask
python-chess
requests
pytest
